requirements.txt
# ============================================================
# requirements.txt — All Python dependencies
# Install with: pip install -r requirements.txt
# ============================================================

# Groq — Fast LLM API (free tier available at console.groq.com)
groq>=0.9.0

# Tavily — Web search API built for AI agents (free tier at app.tavily.com)
tavily-python>=0.3.0

# ChromaDB — Local vector database for semantic memory
chromadb>=0.5.0

# BeautifulSoup — HTML parsing for the read_url tool
beautifulsoup4>=4.12.0

# Requests — HTTP client for fetching URLs
requests>=2.31.0

# lxml — Faster HTML parser (used by BeautifulSoup)
lxml>=5.0.0
